Add clean PDF endpoint for patient medical certificate
